대회 정보 API 리팩토링?
