--- back-end/src/main/java/org/antop/billiardslove/api/ContestModifyRequest.java
@@ -1,9 +1,0 @@
-package org.antop.billiardslove.api;
-
-/**
- * 대회 수정 요청
- *
- * @author jammini
- */
-public class ContestModifyRequest extends ContestRegistrationRequest {
-}
